Rutas protegidas backend

+++ src/app/api/categories/[id]/route.js
@@ -1,10 +1,17 @@
 import { PrismaClient } from '@prisma/client';
+import withAuth from '../../auth/withAuth';
 
 const prisma = new PrismaClient();
 
 export async function GET(request, { params }) {
     try {
-        console.log(params)
+        const verifyUser = await withAuth(request);
+
+        if (verifyUser.status === 400) {
+            const json = await verifyUser.json();
+            return NextResponse.json(json);
+        }
+
         const { id } = params;
         const gender = await prisma.categories.findUnique({
             where: { id: parseInt(id) }
@@ -29,6 +36,14 @@ export async function GET(request, { params }) {
 
 export async function PUT(request, { params }) {
     try {
+
+        const verifyUser = await withAuth(request);
+
+        if (verifyUser.status === 400) {
+            const json = await verifyUser.json();
+            return NextResponse.json(json);
+        }
+
         const { name } = await request.json();
         const { id } = params;
         const updatedGender = await prisma.categories.update({
@@ -49,6 +64,14 @@ export async function PUT(request, { params }) {
 
 export async function DELETE(request, { params }) {
     try {
+
+        const verifyUser = await withAuth(request);
+
+        if (verifyUser.status === 400) {
+            const json = await verifyUser.json();
+            return NextResponse.json(json);
+        }
+
         const { id } = params;
         const deletedGender = await prisma.categories.delete({
             where: { id: parseInt(id) }
