+++ src/app/api/races/[id]/route.js
@@ -1,9 +1,18 @@
 import { PrismaClient } from '@prisma/client';
+import withAuth from '../../auth/withAuth';
 
 const prisma = new PrismaClient();
 
 export async function GET(request, { params }) {
     try {
+
+        const verifyUser = await withAuth(request);
+
+        if (verifyUser.status === 400) {
+            const json = await verifyUser.json();
+            return NextResponse.json(json);
+        }
+
         console.log(params)
         const { id } = params;
         const race = await prisma.races.findUnique({
@@ -29,6 +38,14 @@ export async function GET(request, { params }) {
 
 export async function PUT(request, { params }) {
     try {
+
+        const verifyUser = await withAuth(request);
+
+        if (verifyUser.status === 400) {
+            const json = await verifyUser.json();
+            return NextResponse.json(json);
+        }
+
         const { name } = await request.json();
         const { id } = params;
         const updatedGender = await prisma.races.update({
@@ -49,6 +66,14 @@ export async function PUT(request, { params }) {
 
 export async function DELETE(request, { params }) {
     try {
+
+        const verifyUser = await withAuth(request);
+
+        if (verifyUser.status === 400 ) {
+            const json = await verifyUser.json();
+            return NextResponse.json(json);
+        }
+
         const { id } = params;
         const deletedGender = await prisma.races.delete({
             where: { id: parseInt(id) }
